third_party/blink/web_tests/http/tests/images/drag-drop-image-filename-expected.txt

CONSOLE MESSAGE: Test image loaded.
CONSOLE MESSAGE: Drop event fired.
Filename being dragged: load-and-stall.gif
